Floating Point Error Fix And Initial Reformatting
Fixed the major floating point error in score and migrated score related functions from main.
Changed users.hasCmdAccess to take in author instead of member.
I intend on using the addCommand() function in main to clean it up. Most chat commands such as send and pin will be moved to altcommands in the future