New Score Commands and Fixes
It now ignores all *s while parsing messages for results (This causes first place to always have a score of NaN).

$score info - shows some of the internal variables and links to the score message

$score setfeed <'results' or 'score'> - sets the desired channels to parse for results

Parsing results is now handled in the score module making it 1 line in the main module.

Something to add in the future: Allow for multiple scores so a single bot may be used in multiple servers at once