Self-Update & Minor Fixes
Added start.js which runs the bot as a child process. This allows for code to be ran after the bot process ends. Upon receiving a custom exit code with a command the bot *should* download github files and replace everything within the 'bot-files' folder.

Also New But Untested/Unfinished:
-adding and checking bans in users
-adding bans and DQs in comp

Minor Changes:
-Moved reaction commands from main to miscfuncs
-Deleted unused functions in save
-Correct spelling of celcius and farenheight lmfao