Daily Quiz Module and Touch Ups
whosthatpokemon.js - the latest module that does what the name implies! Each quiz needs to be setup manually before it takes place. Goals: Create a database that it will randomly select from if no quiz is setup manually. Allow it to run in multiple servers at once.

save.js - togglable option to see when files are saved/loaded for debugging

users.js - Passing the author of a message as the string "BOT" is now a shortcut to allow

announcement.js - wrote a better wrapper for delaying internal functions. Can now see all announcements by passing "admin" as an argument to $aclist (internals used to be hidden)

Added eris-additions to await messages sent in channels